fix(harness_k3s): prevent volume leak on teardown
The current implementation of the k3s harness has a tendency to leak config
volumes at the end of each run. This change detaches the creation of the volume
from the creation of the container and includes a clean up step that deletes
the config volume created by the harness itself, preventing that specific leak.

Signed-off-by: Mauren Berti <mauren.berti@chainguard.dev>