use a BaseHarnessResource for removing harness boilerplate (#143)
refactors most of the harness_resource logic into the
`BaseHarnessResource` struct. this is in prep of adding more harnesses,
which previously involved quite a bit of copy/paste and indirection.

this new approach leverages more of the `BaseHarnessResource` to handle
the bulk of the logic, now wrapped in `do(...)`. this also centralizes
the harness creation logic so we can use `Create()` and `Update()`
properly.

The pattern for creating a new harness is to now:

1. create a `imagetest_harness_*.go` resource
2. create an impl of `itypes.Harness` (this will change in an immediate
follow up PR) from the resource schema
3. feed the harness impl into the `BaseHarnessResource.create()`

all the general harness logic is now squirreled away in
`BaseHarnessResource`.