More logical features (#145)
This PR cleans up a buncha dead code around the harness and feature
impls. In the first iteration of this provider, we didn't have the
`data.inventory` to reconcile harnesses->features and needed to compute
this ourselves. Thats why the existing impl of `harness.Base` used
channels to keep things in sync.

Ever since we've had `data.inventory`, the need for concurrency
management of features as they relate to harnesses is no longer needed.
We have a clear "stopping point" when the last feature referencing a
harness is completed and we can start the teardown. This means the
`harness.Base` impl is no longer needed.

This PR removes the `harness.Base` impl and replaces it with a simpler
`Harness` interface:

```go
type Harness interface {
	Create(context.Context) error
	Destroy(context.Context) error
	Run(context.Context, Command) error
}
```

this impl is moved into the `internal/harness` package (from
`internal/harnesses`). The rename is to prepare for the next PR, which
starts implementing new `internal/provider/harnesses` resources.

Additionally, this PR simplifies the `Feature` ergonomics. In a previous
life this provider was built around an `env` manager, and the `Feature`
impl was abstracted to support this. We've mostly moved away from the
`env` manager and as a result can simplify the creation of `Features`.
The new impl should be a lot clear to read, and actually has some tests
to validate things are WAI. This removes a lot of the boilerplate around
the `Feature.Builder`, and simplifies how the `feature_resource` is
created.

None of this should be user facing changes, its all just cleanup to make
this codebase more maintainable and prep for more harnesses! As such,
none of the existing e2e tests are changed