support update/destroy lifecycles with state (#146)
Previously, if we had any pre-existing state (ie anything but `inmem`
backend was used), this provider wouldn't work because we weren't
accounting for the `Update()` rpc's or the legs where state isn't nil.
The provider would fail with pretty cryptic inventory errors.

This uses the refactoring done in #143 and #145 to support the
`Update()` rpc, as well as handle the `Destroy()` rpc by
short-circuiting the `ModifyPlan` when we're in a destroy.

We don't need this yet, but will if/when we move to opentofu and start
using state.

Importantly, this doesn't actually change the logic if state exists, it
simply makes the provider work. All the logic for a `Update()` is the
same as it would be for a `Create()`, given our use case, "state" in
this sense just means a state file, no existing state is preserved from
one run to another.