refactors the usage of docker into a more cohesive client (#148)
This refactors and consolidates our use of the docker sdk behind a
`internal/docker` pkg. Previously we kind of repurposed
`internal/containers/provider`, which was a relic of an older time.

The new `internal/docker` pkg is more thought out to expose the bits we
need behind a tested api, each instance of `docker.New()` is now backed
by an instance specific `*client.Client`, which sets the stage for
future PRs where we will add support for multiple docker backends. The
new `docker` instance is now structured around a `Start()` command,
where each instance can be responsible for starting multiple containers
independently (the old method was 1:1 with the containers it created).

Using this new client, this PR does the following:

- Implements the existing `internal/harness/docker` with
`internal/docker`
- Implements the existing `internal/harness/k3s` with `internal/docker`

Similar to the other refactoring/cleanup PRs #143 and #145, this has no
end user changes, and I'm _heavily_ relying on the existing set of
integration tests we have! The bulk of this PR is to set the stage for
the future PRs which will add support for alternate docker hosts and
remote harnesses.