Connectors (#152)
An ambition of this test harness is to be able to support running
`steps` against arbitrarily complex test fixtures. In the context of
`imagetest`, which is designed exclusively to support testing Chainguard
_images_, a `step` is something that runs in or against a container.

This PR addresses those ambitions with two additions:

- `imagetest_harness_pterraform`: An arbitrary harness that points to an
existing `terraform` source directory
- `connectors`: An interface for running `steps` on an arbitrary_ish_
container runtimes

To address the elephant in the room, yes, this is running terraform in
terraform, yes, it is nasty in some aspects. Original implementations of
this were scoped to supporting various test fixtures, such as remote GKE
clusters, or remote GCE instances with GPUs attached. After going down
those rabbit holes with the `google-cloud-sdk`, I always resolved to
"this can just be done better with terraform". This is an attempt at
that.

## Why not just use `terraform`

`imagetest` exists because our existing release flow is scoped to a
singular `terraform apply`. no `destroy`, no `update`, and no state.
This limits the lifecycle and dag that we can hook in to. Image if we
just embedded the terraform required for spinning up GKE clusters into
the existing workflow. How would we handle teardown? How would we handle
state? We wouldn't, we'd have to go down the path of orchestrating our
existing terraform into independent operations.

In contrast, this PR takes a different approach and leverages
`imagetest`s existing hackiness around the single `terraform apply`
lifecycle (the `inventory`) to lightly wrap and isolate a `terraform`
invocation into a resource: `imagetest_harness_pterraform`. In
combination with `connectors` (adopted from terraform's own
`remote-exec` connectors), this provides support for arbitrary test
fixtures that simply depend on a valid `connection` output. The
`pterraform` source files are all deconflicted and fully integrated into
the single `terraform apply` lifecycle as the rest of harnesses.

For development, a `IMAGETEST_WORKSPACE` variable is added, which
provides a dev escape hatch to persist the `pterraform` state to a known
directory. In combination with the existing `IMAGETEST_SKIP_TEARDOWN`,
this provides an easier to use dev flow for those pesky test fixtures
that take _ages_ to provision (looking at you EKS).

## What does this do

This wraps a full `terraform apply` + `terraform destroy` lifecycle on
an arbitrary source path into a single `imagetest_harness_pterraform`,
and provides 2 `connectors` for interacting with the `output` of the
source in a structured way.

The `connectors` are similar to how `providers` work with `remote-exec`,
only coded up with `imagetest`.

### Example

To help explain, take a remote cloud cluster: `gke`. The following
_outer_ terraform would look something like this:

```hcl
# The "outer" terraform
data "imagetest_inventory" "this" {}

resource "imagetest_harness_pterraform" "foo" {
  name      = "foo"
  inventory = data.imagetest_inventory.this

  path = "${path.module}/gke"
}

resource "imagetest_feature" "foo" {
  name    = "Footure"
  harness = imagetest_harness_pterraform.foo
  steps = [
    {
      name = "Interact with the harness cluster"
      cmd  = "kubectl get po -A"
    },
  ]
}
```

Meanwhile, the "inner" pterraform might look something like:

```hcl
# ${path.module}/gke
terraform { required_providers { google = {} } }

resource "google_container_cluster" "this" {
  name = var.name
  ...
}

module "gke_auth" {
  source = "terraform-google-modules/kubernetes-engine/google//modules/auth"
  ...
}

output "connection" {
  sensitive = true
  value = {
    kubernetes = {
      kubeconfig = module.gke_auth.kubeconfig_raw
    }
  }
}
```

In the above, the `imagetest_feature` would use the
`output.connection.kubernetes` to wire in the connection for the
`imagetest_harness_pterraform.foo` resource, and the subsequent steps
would be run in a sandbox pod within the created GKE cluster.