retry various network create errors (#161)
"hardcode" a capped backoff retry for network create requests.

while the daemon does its best to find a non-conflicting cidr range to
create a network on, when multiple concurrent requests are trying to
create a network (such as when multiple harnesses are being created),
theres a relatively common case where conflicts happen.

a "dumb" retry like this seems to be common place in other solutions
like
[`kind`](https://github.com/kubernetes-sigs/kind/blob/0f1c56884daf22bd4d1559f9b955c26afaaeefb5/pkg/cluster/internal/providers/docker/network.go#L84),
so this approach just adds a non-configurable retry loop with a
reasonable backoff and cap. we could get more fancy and predetermine a
free cidr space during a locked op, but that feels unnecessary for now.