create a "sandbox" interface (#169)
this PR is entirely internal, and only touches
`imagetest_harness_pterraform`, which is yet to be used downstream.

this creates a new `Sandbox` interface that sets us up for a future
world where the test sandboxes run on k8s pods instead of within docker
containers. most of this code is repurposed from `pterraform`, which
already leveraged this model, this merely moves it into a `Sandbox`
interface.

this means that in the future, the test sandboxes for something like the
k3s in docker harness will be run as a cluster privileged pod.