feat: add IMAGETEST_SKIP_ALL for skipping tests at apply time (#193)
Quick and dirty addition of `IMAGETEST_SKIP_ALL`. I plan on following
this up with some additional functionality and cleaning things up.

I noticed that currently the skip implementation in the harness will
skip all features associated with a harness, even if only a single
feature matches the skip criteria. Perhaps this won't matter if we
solidify a 1:1 relationship between harnesses and feature tests, but the
problem will probably still exist if we do something like
`imagetest_suite` which allows multiple `tests` in one block

Signed-off-by: alex weidner <alex.weidner@chainguard.dev>