refactor the inventory impl to avoid racing (#202)
this simplifies and refactors the `inventory` package and its usage to
fix the race condition currently plaguing HEAD.

this implementation simply leans in to the file backend and removes the
interface. and is more aggressive with locks. this means we'll likely
pay _some_ performance penalty with the more aggressive locking, but
we'll be a lot safer against races. given the scale (# of features
relative to a harness), this seems like a worthwhile tradeoff

this isn't intended to be a perfectly highly performant solution, rather
a solution that will buy us enough time for the mono resource approach
coming soon, which voids `inventory` entirely.