add entrypoint binary (#281)
Initial work for the `entrypoint` binary, a small little tool to be the
entrypoint for future imagetest containers built using `tests_resource`.
Each test container will use this entrypoint alongside a user defined
`cmd`, for example:

```hcl
resource "imagetest_tests" "foo" {
  ...

  tests = [
    {
      name    = "helm"
      image   = module.bash_sandbox.image_ref
      content = [{ source = path.module }]
      cmd     = "/imagetest/test.sh"
    }
  ]
}
```

Will produce the following test image:

```
"config": {
    "Cmd": [
      "/imagetest/test.sh"
    ],
    "Entrypoint": [
      "/ko-app/entrypoint",
      "/var/run/ko/entrypoint-wrapper.sh"
    ],
    ...
}
```

This is really similar to prow and tektons entrypoint, in that it runs
as PID 1, forks logs, and handles signalling. The difference with this
one is it doesn't use prow's markers or tekton's step files, and instead
uses a stupid simple health endpoint for querying the status of the
wrapped process.

This means the runtime (ie k8s or docker) can poll the container using
this entrypoint for status on the wrapped process, and react
accordingly. Since we don't need to support multiple steps, we don't
pull in the same logic as tektons entrypoint, and since we use a
healthcheck endpoint instead of markers, we don't pull in prow's
entrypoint.

This also includes the `ko` publishing steps, which publish the
entypoint image and its wrapper sh script as an image on this repo which
`imagetest` (the provider) will ultimately use to append to the test
sandboxes.