add tests_resource (#272)
this PR adds a new resource `imagetest_tests`, which moving forward will
be the _sole_ resource required for authoring image tests with this
provider.

I did my best to document inline in the provider, but the high level is
this canaries the rollover from the "old" inventory/harness/feature
resources into a single `tests` resource.

### `imagetest_tests`

Each `tests` resource is responsible for defining a "driver", which
determines where and how each `test` is run. This is similar to a
`harness`. The difference here is all the setup/run/teardown logic is
now handled in the resource, so the hackiness/buginess of provider
scoped state is gone.

### Tests as images

The other major change this introduces is to treat the tests as a
dynamically built containers that run to completion (in their respective
`drivers`). These test containers are dynamically built in the provider
depending on the driver.

Every test image has access to `$IMAGES`, which is populated via the
images OCI config, and contains the images under test split into their
constituent parts.

Additionally, users can now define their "base" image for the test, and
the provider will append the desired test source and `entrypoint` layers
to build a "full" test image.

### `entrypoint`

This PR also uses the new entrypoint binary (#281), which, similar to
tekton/prow, will wrap the user defined `CMD` and handle signalling, log
teeing, and things like test debug mode.