wire in LDFLAGS for pinning entrypoint image at build time (#289)
currently the entrypoint image is floated at whatever :latest is at
build time. this means we currently need a release for entrypoint first,
before releasing for the provider.

this wires in the ko build output to pin the entrypoint image to the
digest at provider build time, allowing everything to be released at
once.

this also wires in presubmit to properly use the temporary image for
tests. presubmit uses ttl.sh so we can avoid wiring repo auth in
presubmit