plumb through exec instead of assuming /bin/sh (#290)
instead of assuming `CMD` is a shell script, just use `exec`.

this means we can support any valid `CMD`, with the downside that shell
scripts will now need to set their own setopts. we trade some
transparency for more compatibility and control.

this also cleans up the scripts logging a bit, and adds a `socat`
registry proxy to support local registries for the dind driver. the
`socat` works almost identically to how containerd mirroring is done
with the k3sindocker driver.