remove unused slog-multi (#294)
`slog-multi` was added a while ago to attempt to prepare to tee logs to
multiple sinks, with the advent of `imagetest_tests` and other log
collectors, this is no longer needed.

we also never really got the location offset right, so all messages were
coming from the wrong (slog-multi) call site instead. this fixes that:

```
2025-02-05T10:32:17.609-0500 [INFO]  provider.terraform-provider-imagetest: }: @caller=/Users/wolf/dev/gh/chainguard/terraform-provider-imagetest/internal/drivers/docker_in_docker/driver.go:147 @module=imagetest.imagetest test_name=sample test_ref=localhost:5555/foo/imagetest@sha256:d612f122ed1b576cbbb2aa3a78bdacf1892868e05538c331a6f83ff8deae39ad timestamp=2025-02-05T10:32:17.609-0500
```