deconflict dind address pool (#313)
turns out on certain environments (dind in dind on linux 🫠), the default
address pool for the innerinner daemon (`172.0.0.1/16`) will overlap
with the inner daemon (also on `172.0.0.1/16`).

since they're both bridge networks, this usually doesn't matter UNLESS
we're relying on a default route to the host gateway existing for host
traffic, which we do for accessing a local registry via docker's magic
dns.

this sets a default with a low(er) network space (`/24`) on a likely
non-overlapping cidr space so that when `host.docker.internal` resolves,
it always resolves to the hosts gateway (and not the inner daemon's
gateway).