use a registry proxy builtin to the entrypoint instead of socat (#321)
replace the existing naive `socat` tcp proxy with something _ever so
slightly_ smarter. and hoist all the existing sh logic for determining
local registries into the provider itself.

also ensure that `*.localhost` is added as an `ExtraHost` in docker,
such that on linux environments that do not autoresolve `*.localhost` to
`localhost`, that requests can find the proxy