support multiple pause states (#333)
the `imagetest_tests` resource executes pods/containers to completion,
and uses an `entrypoint` to "pause" execution on failures.

to support keeping the containers running on successful completions
(without a non-zero exit code), we have to create one ourselves, and
then plumb it up through to the runtime (docker/k8s) so the client can
determine whether to return a failure when we're in a pause-with-error
state, or to return successfully when we're in a pause-without-error
state.

this also cleans up some logging