standardize pod run error messages (#358)
it was getting confusing consolidating all the error messages across the
different pod run failure modes, so this just standardizes on a
`PodRunError`.

the goal for this is to ensure pod logs are ~always surfaced to the
provider for printing as part of the final error message, and exit codes
are always surfaced when available.