bundle the artifacts during finalize (#357)
landing
[this](https://github.com/chainguard-dev/terraform-provider-imagetest/pull/356)
in two parts.

this is the first part, which configures the `entrypoint` to bundle up
the `/tmp/imagetest/artifacts` dir during its `finalize` phase.

the [follow up
PR](https://github.com/chainguard-dev/terraform-provider-imagetest/pull/356)
will implement the client (tf provider) logic for fetching this
artifact.

one notably change here is we're removing the additional log file
strictly for stderr. this is unused since practically speaking, stderr
is part of the `process.log`, and easier to read when its interleaved
with stdout