AWS Lambda test resource (#353)
This adds a new test resource:

```
resource "imagetest_tests_lambda" "foo" {
  name      = "foo"
  image_ref = "12345.dkr.ecr.us-west-2.amazonaws.com/blah@sha256:07a99..."
}
```

This will deploy the image to a new Lambda function (using creds/config
in your local AWS profile), wait for it to become ready, invoke it, and
check the response was successful. It then deletes the function, unless
you've told it not to.

The intention is to be able to test Lambda images on real Lambda
deployments, like we have for EKS today. To use it, we'd take a base
image we produce, add an application layer on top of it, and deploy and
invoke it, to test that the base image works as expected.

To run the acceptance test:

```
IMAGETEST_LAMBDA_TEST_IMAGE_REF=12345.dkr.ecr.us-west-2.amazonaws.com/foo@sha256:07a99c... \
IMAGETEST_LAMBDA_ROLE=arn:aws:iam::12345:role/lambda-ex \
TF_ACC=1 \
  go test -tags=lambda ./internal/provider/... -count=1 -v -run=Lambda -timeout=5m
```

This takes about 10-15 seconds and passes consistently.

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>