refactor pod.Run and fix a race (#378)
the monitoring logic had gotten waaaaaay too complex with waaaaaay not
enough coverage. there is also a race where the termination event may be
processed before/after the pod completion event is processed, in which
case a successful termination event would be incorrectly marked as a
failure.

so this refactors things to be easier to read/maintain, and adds some
coverage around all the various completion states.