k3s_in_docker: Rename context to cluster name (#387)
This commit renames the default context in the k3s cluster to be the
name of the cluster itself.  This will make it easier for people to
interact with the test cluster(s) from their host machine (watch this
space).  We are not removing the `default` context so as not to break
any existing tooling.

---------

Signed-off-by: Eoghan <eoghan@speleo.dev>
Co-authored-by: Josh Wolf <josh@wolfs.io>