feat(internal/drivers/ec2): create default instance profile
The user should not be required to specify an IAM instance profile
for the EC2 instance the tests are run on. By default, the EC2
instance should be authorized to pull the just build and published
image to ECR transparently for the end-user of the provider.
For this reason, this commit adds the logic to setup the default
IAM for the EC2 instance so that it is launched with the necessary
permissions in AWS to read-only from ECR.

The policy attached to the IAM role is AmazonEC2ContainerRegistryReadOnly,
and the principle trusted to assume the role is exactly the EC2
service principle.

This will allow the image pushed to ECR to be used for testing
. Optionally, the user can still customize the instance profile
in case different set of permissions is required for testing.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>