Configure IAM for EC2 driver to pull from ECR (#544)
The EC2 instance should be authorized to pull the just built and
published image to ECR transparently for the end-user of the provider.
For this reason, this commit adds the logic to setup the default
IAM for the EC2 instance so that it is launched with the necessary
permissions in AWS to read-only from ECR.

The policy attached to the IAM role is
AmazonEC2ContainerRegistryReadOnly,
and the principle trusted to assume the role is exactly the EC2
service principle.

This will allow the image pushed to ECR to be used for testing by
the EC2 driver.

The IAM resources are pushed to the cleanup stack so that teardown
is  managed properly and in the correct order.

Optionally, the user can customize the instance profile
in case different set of permissions is required for testing.

---------

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>
Signed-off-by: maxgio92 <massimiliano.giovagnoli@chainguard.dev>