Avoid health check racing container exit
Sometimes, we see tests fail with messages like `docker-in-docker test failed: container exited with non-zero exit code: -1: OCI runtime exec failed: exec failed: cannot exec in a stopped container: unknown`. We think the exec here is referring to the container health check which docker wants to execute after the container has already stopped.

This fixes that race by only considering health state if the container is in a running state. It's moot otherwise.