ec2: refactor the driver (#635)
this is a rewrite of the `ec2` driver.

- slims down the required infra owned by the driver
- the prior driver created "full" VPCs (NAT/EIP) per test, which
included the globally scarce EIPs, and wasn't properly cleaning things
up in all cases, which led to quickly exhausting an accounts allocated
EIPs
- now it expects a pre-existing VPC, and just uses public instances
locked to the executors public IP
- removes the gpu specific userdata/postrun scripts
- this makes the driver more generic, and lets the caller handle whether
they want to mount GPUs or not. that also means the caller is
responsible for any gpu setup, but that's now more easily handled with
userdata and better connection checking across things like reboots
- more resilient cleanups
- all the resources this driver owns are now typed, with each one being
responsible for its own cleanup. this makes it easier in the code to
follow what the dependencies are, and who is responsible for cleaning up
what
- aws sdk heleprs are used to properly wait for dependents to be cleaned
up with appropriate polling/timeouts
- all resources are tagged with attributions back to owner, and other
available env details