feat(eks_with_eksctl): use global timeout for eksctl operations
The imagetest_test module has a top-level `timeout` option that controls
the maximum time the tests may run.

This setting is global for the module and is not downstreamed to any of
the underlying drivers. This means that, when users set a timeout on the
imagetest_tests resource, it only affects the overall Go context but not
eksctl's internal operations, which defaulted to 25 minutes. This means
eksctl can timeout before the user's configured timeout was reached.

The eks_with_eksctl driver uses the eksctl CLI to manage clusters on
AWS. The default timeout for its operations is 25 minutes and it allows
overriding that with the --timeout flag.

This change makes it possible to pass the top-level timeout to the
eks_with_eksctl driver. Now the timeout value flows through to eksctl
commands via the --timeout flag, ensuring eksctl respects the user's
timeout setting for cluster creation, node group operations, and other
long-running tasks.