eks: plumb static auth through (#653)
this is a follow up to this comment:
https://github.com/chainguard-dev/terraform-provider-imagetest/pull/651#discussion_r2606913043

going for it now because this also adds support for plumbing through
static auth for the `ekswitheksctl` driver, and adding yet another named
argument to the `LoadDriver` felt like the tipping point!

this leverages the same static auth "trick" that the k3s driver uses, to
create a `dockercredential` secret with the resolved auth. the
difference for the eks driver is that it only uses the tests `repo`
override (which is usually plumbed through as an `ecr` repo)